Handle changed preferred_address situations.
So our primary mail (@scs.sovereignit.de) might not be the user's
preferred address, so we need to be a lot more clever in checking
subscription status etc.
Also even if we are not subscribed yet, the user might already exist
with any of his mail addresses, so search for it, add missing mails
and then ensure the subscription status is correct:
* Once as member (preferrably with the preferred address)
* All other addresses as non-members

To simplify the code, we also move initialize() and userManager =
out of the functions.

Signed-off-by: Kurt Garloff <kurt@garloff.de>